Add listener for mouse click on button and refactoring
