+++ src/main/java/org/example/Main.java
@@ -1,6 +1,6 @@
 package org.example;
 
-import org.example.client.Window;
+import org.example.client.ui.screen.Window;
 
 public class Main {
     public static void main(String[] args) {
