+++ src/main/java/org/example/client/ui/listener/MouseListener.java
@@ -1,4 +1,4 @@
-package org.example.client;
+package org.example.client.ui.listener;
 
 import static org.lwjgl.glfw.GLFW.GLFW_PRESS;
 import static org.lwjgl.glfw.GLFW.GLFW_RELEASE;
