+++ src/main/java/org/example/client/ui/screen/ServerSelectorScene.java
@@ -1,7 +1,9 @@
-package org.example.client;
+package org.example.client.ui.screen;
 
 
-import org.example.client.model.Rectangle;
+import org.example.client.ui.listener.KeyListener;
+import org.example.client.ui.listener.MouseListener;
+import org.example.client.ui.model.Rectangle;
 import org.example.client.structure.Vector2;
 import org.example.client.structure.Vector3;
 
@@ -25,17 +27,20 @@ public class ServerSelectorScene extends Scene {
     }
 
     private Rectangle createButton( float x, float y, int width, int height) {
-        return new Rectangle(Vector2.of(x, y), width, height, Vector3.of(0.0f, 0.0f, 1.0f));
+        return new Rectangle(Vector2.of(x, y), width, height, Vector3.of(0.0f, 0.0f, 0.0f));
     }
 
     @Override
     public void update(float dt) {
         button.draw();
 
-        System.out.println((1.0f / dt) + " FPS");
+        //System.out.println((1.0f / dt) + " FPS");
 
-        if (!changingScene && KeyListener.isKeyPressed(KeyEvent.VK_SPACE)) {
-            changingScene = true;
+        if(MouseListener.mouseButtonDown(0)) {
+            if(button.contains(MouseListener.getX(), MouseListener.getY())) {
+                System.out.println("Button pressed");
+                changingScene = true;
+            }
         }
 
         if (changingScene && timeToChangeScene > 0) {
